Work in progress

+++ src/scriptengine.cpp
@@ -932,7 +932,7 @@ void framework::scriptengine::run() {
   engine->add_loopable(std::make_shared<lua_loopable>(lua, loop));
 
   const auto end = SDL_GetPerformanceCounter();
-  const auto elapsed = (end - start) * 1000.0 / static_cast<double>(SDL_GetPerformanceFrequency());
+  const auto elapsed = static_cast<double>(end - start) * 1000.0 / static_cast<double>(SDL_GetPerformanceFrequency());
   fmt::println("boot time {:.3f}ms", elapsed);
 
   engine->run();
